removes orig files
What is an ORIG file?

Files that use the .orig file extension usually contain backup files. The .orig file extension is used to notate that a file is an original version of the file when a revised or updated copy of the file is made and the original file is kept for backup purposes.